validation token user or admin

+++ Back/src/courseGroup/course/deleteCourseByName.js
@@ -1,4 +1,5 @@
 const CourseService = require('../../common/services/CourseService')
+const tokenVerification = require('../../../utils/tokenVerification');
 const httpResponse = require('../../../utils/httpResponse');
 const logger = require('../../../utils/logger');
 
@@ -7,11 +8,17 @@ const deleteCourse = async (event) => {
     logger.info('Curso obtenida', curso);
     try {
         const courseService = new CourseService()
-        
+
+        const user = tokenVerification(event);
+        if (!user || user.scope !== 'administrador') {
+            logger.error('Usuario no autorizado para borrar cursos');
+            return httpResponse.unauthorized(new Error('No tienes permiso para borrar cursos'))(event.requestContext.path);
+        }
+
         const res = await courseService.deleteEntity(curso);
 
         logger.info('curso eliminado correctamente', res);
-        return httpResponse.ok({message: 'Curso eliminado correctamente'})
+        return httpResponse.ok({ message: 'Curso eliminado correctamente' })
     } catch (error) {
         logger.error('Error eliminando curso', error);
         return httpResponse.badRequest(new Error('Error al eliminar curso: ' + error.message))(event.requestContext.path);
